+++ Fuchas/Kernel/Startup/00_luaapi.lua
@@ -111,103 +111,6 @@ function table.getn(table)
 end
 table.maxn = table.getn
 
--- Obsolete I/O methods
-function io.fromu16(x)
-	local b1=string.char(x%256) x=(x-x%256)/256
-	local b2=string.char(x%256)
-	return {b1, b2}
-end
-
-function io.fromu32(x)
-	local b1=string.char(x%256) x=(x-x%256)/256
-	local b2=string.char(x%256) x=(x-x%256)/256
-	local b3=string.char(x%256) x=(x-x%256)/256
-	local b4=string.char(x%256)
-	return {b1, b2, b3, b4}
-end
-
-function io.tou16(arr, off)
-	local v1 = arr[off + 1]
-	local v2 = arr[off]
-	return v1 + (v2*256)
-end
-
-function io.tou32(arr, off)
-	local v1 = io.tou16(arr, off + 2)
-	local v2 = io.tou16(arr, off)
-	return v1 + (v2*65536)
-end
-
--- New I/O methods
-
--- To unsigned number (max 32-bit)
-function io.tounum(number, count, littleEndian)
-	local data = {}
-	
-	if count > 4 then
-		error("lua bit32 only supports 32-bit numbers")
-	end
-	
-	if littleEndian then
-		local i = count
-		while i > 0 do
-			data[i] = bit32.band(number, 0x000000FF)
-			number = bit32.rshift(number, 8)
-			i = i - 1
-		end
-	else
-		local i = 1
-		while i < count+1 do
-			data[i] = bit32.band(number, 0x000000FF)
-			number = bit32.rshift(number, 8)
-			i = i + 1
-		end
-	end
-	return data
-end
-
--- From unsigned number (max 32-bit)
-function io.fromunum(data, littleEndian, count)
-	count = count or 0
-	if count == 0 then
-		if type(data) == "string" then
-			count = data:len()
-		else
-			count = #data
-		end
-	end
-	
-	if count > 4 then
-		error("lua bit32 only supports 32-bit numbers")
-	end
-	
-	if count == 1 then
-		if data then
-			return string.byte(data)
-		else
-			return nil
-		end
-	else
-		-- use 4 bytes max as Lua's bit32 scale the number between [0, 2^32-1] which makes the number impossible to
-		-- go beyond ‭4,294,967,295‬
-		local bytes, result = {string.byte(data or "\x00", 1, 4)}, 0
-		if littleEndian then
-			local i = #bytes -- just do it in inverse order
-			while i > 0 do
-				result = bit32.bor(bit32.lshift(result, 8), bytes[i])
-				i = i - 1
-			end
-		else
-			local i = 1
-			while i < #bytes do
-				result = bit32.bor(bit32.lshift(result, 8), bytes[i])
-				i = i + 1
-			end
-		end
-		return result
-	end
-end
-
 -- Convenient Lua extensions
 
 function try(func)
